+++ src/main/java/com/tallerwebi/dominio/ServicioHome.java
@@ -1,0 +1,7 @@
+package com.tallerwebi.dominio;
+
+import java.util.List;
+
+public interface ServicioHome {
+    List<Anuncio> getAnuncios();
+}
