+++ src/main/java/com/tallerwebi/dominio/ServicioJugador.java
@@ -3,6 +3,5 @@ package com.tallerwebi.dominio;
 import java.util.List;
 
 public interface ServicioJugador {
-    List<Anuncio> getAnuncios();
     Jugador getJugadorActual();
 }
