+++ src/main/java/com/tallerwebi/presentacion/ControladorHome.java
@@ -1,5 +1,6 @@
 package com.tallerwebi.presentacion;
 
+import com.tallerwebi.dominio.ServicioHome;
 import com.tallerwebi.dominio.ServicioJugador;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -13,17 +14,19 @@ import javax.servlet.http.HttpServletRequest;
 @Controller
 @RequestMapping("/home")
 public class ControladorHome {
+    private final ServicioHome servicioHome;
     private final ServicioJugador servicioJugador;
 
     @Autowired
-    public ControladorHome(ServicioJugador servicioJugador) {
+    public ControladorHome(ServicioHome servicioHome, ServicioJugador servicioJugador) {
+        this.servicioHome = servicioHome;
         this.servicioJugador = servicioJugador;
     }
 
     @GetMapping()
     public ModelAndView getHome(HttpServletRequest request) {
         ModelMap model = new ModelMap();
-        model.addAttribute("anuncios", this.servicioJugador.getAnuncios());
+        model.addAttribute("anuncios", this.servicioHome.getAnuncios());
         model.addAttribute("jugador", this.servicioJugador.getJugadorActual());
         return new ModelAndView("home", model);
     }
